주문 수량 불일치하던 현상 수정 & 소수점 10자리로 제한

+++ src/main/java/bit/bitgroundspring/dto/TradeDto.java
@@ -12,8 +12,8 @@ public class TradeDto {
     private String symbol;
     private String koreanName;
     private OrderType orderType;
-    private Float amount;
-    private Float tradePrice;
+    private Double amount;
+    private Double tradePrice;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
