+++ src/main/java/bit/bitgroundspring/dto/projection/UserAssetProjection.java
@@ -2,6 +2,6 @@ package bit.bitgroundspring.dto.projection;
 
 public interface UserAssetProjection {
     String getSymbol();
-    Float getAmount();
-    Float getAvgPrice();
+    Double getAmount();
+    Double getAvgPrice();
 }
